Init Backend Auth

+++ backend/src/main/java/com/example/backend/db/repository/CottageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.backend.db.model.Cottage;
 
-public interface CottageRepository extends JpaRepository<Cottage, Integer> {
+public interface CottageRepository extends JpaRepository<Cottage, Long> {
     
 }
